add atan2 host benchmark
