+++ dsp/Cargo.toml
@@ -12,7 +12,7 @@ serde = { version = "1.0", features = ["derive"], default-features = false }
 criterion = "0.3"
 
 [[bench]]
-name = "cossin"
+name = "trig"
 harness = false
 
 [features]
